requirements.txt
google==1.9.1
requests==2.8.1
tweepy==3.5.0
xmltodict==0.10.2
pyyaml==3.11
flask==0.10.1
gunicorn==19.4.5
psycopg2==2.5.3
